GeometryImage/geometry_image/requirements.txt
numpy==1.20.3
vtk==9.0.3
Pillow==8.3.1
scipy==1.7.1